pyproject.toml
[tool.poetry]
name = "python_auth"
version = "0.1.0"
description = ""
authors = ["Ravi Bharti <ravi@aganitha.ai>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.82.0"
SQLAlchemy = "^1.4.40"
pydantic = "^1.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
uvicorn = "^0.18.3"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
